requirements.txt
tweepy
pillow